pyproject.toml
[project]
name = "tiny-chaos"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "matplotlib>=3.10.3",
    "pandas>=2.2.3",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
    "ruff>=0.11.9",
    "websockets>=12.0",
]
